+++ awx/ui_next/src/constants.js
@@ -2,7 +2,7 @@
 export const JOB_TYPE_URL_SEGMENTS = {
   job: 'playbook',
   project_update: 'project',
-  system_job: 'system',
+  system_job: 'management',
   inventory_update: 'inventory',
   ad_hoc_command: 'command',
   workflow_job: 'workflow',